requirements.txt
anyio==4.8.0
boto3==1.36.16
botocore==1.36.16
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==20.0
rfc3986==1.5.0
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
